move util functions out of query.go
